requirements.txt
dash==0.39  # The core dash backend
dash-html-components==0.14.0  # HTML components
dash-core-components==0.44.0  # Supercharged components
dash-table==3.6.0  # Interactive DataTable component (new!)
dash-daq==0.1.0  # DAQ components (newly open-sourced!)
dash
dash-renderer
dash-html-components
dash-core-components
plotly
pandas
numpy
happybase
tqdm
pyspark
nltk
kafka-python